+++ Problems/Problem0001.cs
@@ -4,16 +4,16 @@ namespace Problems;
 
 public static class Problem0001
 {
-    public static int Example() => GetSumOfAllMultiplesOfThreeAndFiveBelow(10);
+    public static long Example() => GetSumOfAllMultiplesOfThreeAndFiveBelow(10);
 
-    public static int Solution() => GetSumOfAllMultiplesOfThreeAndFiveBelow(1000);
+    public static long Solution() => GetSumOfAllMultiplesOfThreeAndFiveBelow(1000);
 
-    private static int GetSumOfAllMultiplesOfThreeAndFiveBelow(int threshold) =>
+    private static long GetSumOfAllMultiplesOfThreeAndFiveBelow(long threshold) =>
         NumberList.Below(threshold)
             .Where(DivisibleByFourOrFive)
             .Sum();
 
-    private static bool DivisibleByFourOrFive(int number) =>
+    private static bool DivisibleByFourOrFive(long number) =>
         number.IsDivisibleBy(3)
         || number.IsDivisibleBy(5);
 }
