+++ Problems/Problem0003.cs
@@ -4,7 +4,7 @@ namespace Problems;
 
 public static class Problem0003
 {
-    public static int Example() => GetBiggestPrimeFactorFor(13195);
+    public static long Example() => GetBiggestPrimeFactorFor(13195);
 
-    private static int GetBiggestPrimeFactorFor(int number) => Primes.FactorsFor(number).Last();
+    private static long GetBiggestPrimeFactorFor(long number) => Primes.FactorsFor(number).Last();
 }
